Injection de la publicité dans le controlleur de la notification

+++ database/migrations/2024_03_20_113751_create_rapport_taches_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->collation = 'utf8_unicode_ci';
 
             $table->increments('idRapport');
-            $table->unsignedInteger('idUser');
+            $table->unsignedBigInteger('idUser');
             $table->foreign('idUser')->references('id')->on('users')->onDelete('cascade');
             $table->string('valider');
             $table->text('encour')->nullable();
